+++ packages/manage/src/components/InlangManage.ts
@@ -265,7 +265,7 @@ export class InlangManage extends TwLitElement {
 	}
 
 	async removeLanguageTag(languageTag: string) {
-		if (typeof this.user === "undefined") return
+		if (typeof this.user === "undefined" || this.user === "load") return
 
 		this.languageTags = this.languageTags?.map((tag) => {
 			if (tag.name === languageTag) {
@@ -333,7 +333,7 @@ export class InlangManage extends TwLitElement {
 	async addLanguageTag() {
 		if (this.newLanguageTag === "") return
 
-		if (typeof this.user === "undefined") return
+		if (typeof this.user === "undefined" || this.user === "load") return
 
 		this.newLanguageTagLoading = true
 
@@ -705,7 +705,7 @@ export class InlangManage extends TwLitElement {
 										To access your projects, please enter the URL of your GitHub repository.
 									</p>
 									<div
-										disabled=${this.url.repo}
+										disabled=${this.url?.repo === undefined}
 										class=${`px-1 gap-2 relative max-w-lg z-10 flex items-center w-full border bg-white rounded-lg transition-all relative ${
 											this.url.repo ? "cursor-not-allowed" : ""
 										} ${
@@ -847,8 +847,7 @@ export class InlangManage extends TwLitElement {
 										</a>
 									</div>
 							  </div>`
-							: this.projects !== "load" &&
-							  this.projects !== "no-access" &&
+							: this.projects !== "no-access" &&
 							  // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
 							  this.projects!.length === 0
 							? html`<div class="flex flex-col gap-0.5 mt-4">
@@ -874,7 +873,6 @@ export class InlangManage extends TwLitElement {
 									</div>
 							  </div>`
 							: !this.url.project &&
-							  this.projects !== "load" &&
 							  this.projects !== "no-access" &&
 							  // eslint-disable-next-line @typescript-eslint/no-non-null-assertion
 							  this.projects!.length > 0