+++ packages/paraglide/paraglide-js-adapter-next/build.mjs
@@ -1,17 +1,19 @@
 import { rollup } from "rollup"
-import preserveDirectives from "rollup-preserve-directives"
-import typescript from "@rollup/plugin-typescript"
-import cjs from "@rollup/plugin-commonjs"
-import resolve from "@rollup/plugin-node-resolve"
+import { plugins } from "./plugins.mjs"
 import fs from "node:fs/promises"
 
 //remove the old dist folder
 await fs.rm("./dist", { recursive: true, force: true })
 
 const packageJson = JSON.parse(await fs.readFile("./package.json", "utf-8"))
+const version = packageJson.version
+if (typeof version !== "string") {
+	throw new Error("the version in package.json is not a string")
+}
 const peerDependencies = Object.keys(packageJson.peerDependencies || {})
 const dependencies = Object.keys(packageJson.dependencies || {})
 
+// eslint-disable-next-line no-undef
 // eslint-disable-next-line no-console
 // eslint-disable-next-line no-undef
 console.info(`Building ${packageJson.name} v${packageJson.version}...`)
@@ -25,7 +27,7 @@ const external = [
 ]
 
 const app_build = await rollup({
-	plugins: [typescript({ tsconfig: "./tsconfig.json" }), cjs(), resolve(), preserveDirectives()],
+	plugins,
 	input: {
 		"app/index.server": "src/app/index.server.tsx",
 		"app/index.client": "src/app/index.client.tsx",
@@ -34,7 +36,7 @@ const app_build = await rollup({
 })
 
 const pages_build = await rollup({
-	plugins: [typescript({ tsconfig: "./tsconfig.json" }), cjs(), resolve(), preserveDirectives()],
+	plugins,
 	input: {
 		"pages/index": "src/pages/index.tsx",
 	},
@@ -42,11 +44,19 @@ const pages_build = await rollup({
 })
 
 const pluginBuild = await rollup({
-	plugins: [typescript({ tsconfig: "./tsconfig.json" }), cjs(), resolve(), preserveDirectives()],
+	plugins,
 	input: {
 		"plugin/index": "src/plugin/index.ts",
 	},
 	external,
+})
+
+const cliBuild = await rollup({
+	plugins,
+	input: {
+		"cli/index": "src/cli/index.ts",
+	},
+	external: [/^node:/, ...peerDependencies, ...dependencies],
 })
 
 await app_build.write({
@@ -67,5 +77,12 @@ await pluginBuild.write({
 	preserveModules: false,
 	format: "cjs",
 	entryFileNames: "[name].cjs",
+	dir: "dist",
+})
+
+await cliBuild.write({
+	preserveModules: true,
+	format: "esm",
+	entryFileNames: "[name].mjs",
 	dir: "dist",
 })