+++ packages/paraglide/paraglide-js-adapter-next/examples/pages/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @inlang/paraglide-js-adapter-next-example-pages
 
+## 0.2.19
+
+### Patch Changes
+
+- Updated dependencies [e5ce0bd]
+- Updated dependencies [6c08db1]
+  - @inlang/paraglide-js-adapter-next@3.3.0
+
+## 0.2.18
+
+### Patch Changes
+
+- Updated dependencies [0774c1a]
+- Updated dependencies [0774c1a]
+  - @inlang/paraglide-js@1.7.0
+  - @inlang/paraglide-js-adapter-next@3.2.0
+
+## 0.2.17
+
+### Patch Changes
+
+- Updated dependencies [cee4692]
+- Updated dependencies [4b631aa]
+- Updated dependencies [3c7a87c]
+- Updated dependencies [ab1fe48]
+  - @inlang/paraglide-js-adapter-next@3.1.1
+  - @inlang/paraglide-js@1.6.2
+
 ## 0.2.16
 
 ### Patch Changes